refactor: move channel-context + Channel/ChannelsConfig out of legacy/
After the transport→channel rename, these two files were the only thing
left in src/legacy/gateway/ and they were never legacy concepts to begin
with — just misplaced. Move:

- src/legacy/gateway/channel-context.ts → src/channels/channel-context.ts
- src/legacy/gateway/types.ts merged into src/channels/types.ts
  (Channel + ChannelsConfig now live alongside ChannelAdapter)

Also drops vestigial start/stop from the Channel interface — those
duplicated ChannelAdapter's lifecycle and were stuffed with empty async
fns at every call site. Channel is now purely the agent-tool callback
surface (react, reply).

src/legacy/gateway/ directory deleted entirely.
